requirements.txt
black
scikit-learn
tensorflow
sqlalchemy
python-binance
pandas
numpy
tqdm
psycopg2-binary
pydot
pydotplus
matplotlib